Feature/leaderboard route (#62)
* added the page.tsx for homepage back since it was accidently deleted since I was trying to just eliminate the changes done for testing the leaderboard component.

* Merge branch 'main' of https://github.com/acmucsd-projects/fa2024-hack-team-2

* Add leaderboard route and enhance user and post models with likes functionality

* Enhance post model to support multiple images and update related routes for image handling

* Add documentation for leaderboard retrieval endpoint

+++ server/routes/post.js
@@ -16,9 +16,10 @@ const express_1 = __importDefault(require("express"));
 const Post_1 = __importDefault(require("../models/Post"));
 const User_1 = require("../models/User");
 const mongoose_1 = __importDefault(require("mongoose"));
-const User_1 = require("../models/User");
-const mongoose_1 = __importDefault(require("mongoose"));
+const multer_1 = __importDefault(require("multer"));
 const router = express_1.default.Router();
+const storage = multer_1.default.memoryStorage();
+const upload = (0, multer_1.default)({ storage: storage });
 /**
  * @route POST /
  * @desc Create a new post
@@ -44,13 +45,33 @@ const router = express_1.default.Router();
  * - 401: Unauthorized (if the user is not authenticated).
  * - 500: Internal server error.
  */
-router.post("/", (req, res) => __awaiter(void 0, void 0, void 0, function* () {
+router.post("/", upload.array("images", 3), (req, res) => __awaiter(void 0, void 0, void 0, function* () {
     if (!req.user) {
         res.status(401).json({ error: "Unauthorized" });
         return;
     }
     try {
-        const { title, product_details, material, brand, cost, numStores, available_stores, image, tags, } = req.body;
+        const author = yield User_1.User.findOne({
+            user_id: req.user.user_id,
+        });
+        if (!author) {
+            res.status(404).json({ error: "User not found" });
+            return;
+        }
+        const { title, product_details, material, brand, cost, numStores, available_stores, tags, } = req.body;
+        const images = req.files
+            ? req.files.map((file) => ({
+                data: file.buffer,
+                contentType: file.mimetype,
+            }))
+            : [];
+        // Debug statements for images
+        console.log(`Number of images uploaded: ${images.length}`);
+        images.forEach((image, index) => {
+            console.log(`Image ${index + 1}:`);
+            console.log(`- Content Type: ${image.contentType}`);
+            console.log(`- Size: ${image.data.length} bytes`);
+        });
         const newPost = new Post_1.default({
             title,
             product_details,
@@ -59,9 +80,8 @@ router.post("/", (req, res) => __awaiter(void 0, void 0, void 0, function* () {
             cost,
             numStores,
             author: req.user.user_id,
-            author: req.user.user_id,
             available_stores,
-            image,
+            images,
             tags,
             date_created: new Date().toLocaleDateString("en-US", {
                 year: "numeric",
@@ -69,8 +89,11 @@ router.post("/", (req, res) => __awaiter(void 0, void 0, void 0, function* () {
                 day: "2-digit",
             }),
         });
+        // adding post to the author's list of posts
+        author.posts.push(new mongoose_1.default.Types.ObjectId(newPost._id));
         const savedPost = yield newPost.save();
-        res.status(201).json(savedPost);
+        yield author.save();
+        res.status(201).json({ message: "Post created successfully", savedPost });
     }
     catch (error) {
         console.error("Error creating post:", error);
@@ -100,7 +123,12 @@ router.get("/", (req, res) => __awaiter(void 0, void 0, void 0, function* () {
             res.status(404).json({ error: "Post not found" });
             return;
         }
-        res.status(200).json(post);
+        // Convert image data to base64-encoded strings
+        const postWithBase64Images = Object.assign(Object.assign({}, post.toObject()), { images: post.images.map(image => ({
+                contentType: image.contentType,
+                data: image.data.toString('base64')
+            })) });
+        res.status(200).json(postWithBase64Images);
     }
     catch (error) {
         console.error("Error fetching post:", error);
@@ -112,7 +140,9 @@ router.get("/", (req, res) => __awaiter(void 0, void 0, void 0, function* () {
  * @desc Delete a post by its _id
  * @access Private
  *
- * This endpoint allows an authenticated user to delete a post.
+ * This endpoint allows an authenticated user to delete a post. The post is deleted from the user's posts
+ * and the totalLikes of the author is decreased by the number of likes the post has.
+ * This allow deletes the post from the liked user's list of liked posts.
  *
  * Request body:
  * - post_id: The ID of the post to be deleted. (required)
@@ -141,6 +171,22 @@ router.delete("/", (req, res) => __awaiter(void 0, void 0, void 0, function* () 
             res.status(401).json({ error: "Unauthorized" });
             return;
         }
+        const author = yield User_1.User.findOne({ user_id: user_id });
+        if (!author) {
+            console.log("Author not found");
+            throw new Error("Author not found");
+        }
+        // Deleting the post from the author's list of posts
+        author.posts = author.posts.filter((postId) => postId.toString() !== post_id.toString());
+        author.totalLikes -= post.likes;
+        // Deleting the post from the liked user's list of liked posts
+        const likedUser = yield User_1.User.find({ user_id: { $in: post.likesList } });
+        likedUser.forEach((likedUser) => __awaiter(void 0, void 0, void 0, function* () {
+            likedUser.liked = likedUser.liked.filter((likedPostId) => likedPostId.toString() !== post_id.toString());
+            yield likedUser.save();
+        }));
+        // Saving the author changes since we are deleting post and decreasing the totalLikes
+        yield author.save();
         yield Post_1.default.deleteOne({ _id: post_id });
         res.status(200).json({ message: "Post deleted successfully" });
     }
@@ -175,7 +221,7 @@ router.delete("/", (req, res) => __awaiter(void 0, void 0, void 0, function* () 
  * - 404: Post not found.
  * - 500: Internal server error.
  */
-router.patch("/", (req, res) => __awaiter(void 0, void 0, void 0, function* () {
+router.patch("/", upload.array("images", 3), (req, res) => __awaiter(void 0, void 0, void 0, function* () {
     if (!req.user) {
         res.status(401).json({ error: "Unauthorized" });
         return;
@@ -214,11 +260,15 @@ router.patch("/", (req, res) => __awaiter(void 0, void 0, void 0, function* () {
         if (available_stores) {
             post.available_stores = available_stores;
         }
-        if (image) {
-            post.image = image;
-        }
         if (tags) {
             post.tags = tags;
+        }
+        if (Array.isArray(req.files) && req.files.length > 0) {
+            const images = req.files.map(file => ({
+                data: file.buffer,
+                contentType: file.mimetype,
+            }));
+            post.images = images;
         }
         yield post.save();
         res.status(200).json({ message: "Post updated successfully", post });
@@ -245,7 +295,12 @@ router.patch("/", (req, res) => __awaiter(void 0, void 0, void 0, function* () {
  */
 router.get("/author", (req, res) => __awaiter(void 0, void 0, void 0, function* () {
     try {
-        const posts = yield Post_1.default.find({ author: req.body.user_id }); // Find all posts by author
+        const user = yield User_1.User.findOne({ user_id: req.body.user_id });
+        if (!user) {
+            res.status(404).json({ error: "User not found" });
+            return;
+        }
+        const posts = yield Post_1.default.find({ _id: { $in: user.posts } });
         res.status(200).json(posts);
     }
     catch (error) {
@@ -261,6 +316,8 @@ router.get("/author", (req, res) => __awaiter(void 0, void 0, void 0, function* 
  * This endpoint allows an authenticated user to like or unlike a post.
  * If the user has already liked the post, it will be unliked.
  * If the user has not liked the post, it will be liked.
+ * The totalLikes of the post and the author will be updated accordingly.
+ * The user's liked posts list will also be updated.
  *
  * Request body:
  * - post_id: The ID of the post to be liked or unliked.
@@ -278,38 +335,45 @@ router.patch("/like", (req, res) => __awaiter(void 0, void 0, void 0, function* 
         return;
     }
     try {
-        console.log("request body:", req.body);
-        const post_id = new mongoose_1.default.Types.ObjectId(req.body.post_id);
+        const post_id = req.body.post_id;
         const user_id = req.user.user_id;
-        console.log("This is the user id: ", user_id);
         // Validate post_id
         if (!mongoose_1.default.Types.ObjectId.isValid(post_id)) {
             res.status(400).json({ error: "Invalid post_id format" });
-            return;
         }
         const post = yield Post_1.default.findById(post_id);
         if (!post) {
             res.status(404).json({ error: "Post not found" });
             return;
         }
-        console.log("Fetched post:", post);
-        const user = yield User_1.User.findOne({ user_id: user_id });
+        const [author, user] = yield Promise.all([
+            User_1.User.findOne({ user_id: post.author }),
+            User_1.User.findOne({ user_id: user_id }),
+        ]);
+        if (!author) {
+            res.status(404).json({ error: "Author not found" });
+            return;
+        }
         if (!user) {
             res.status(404).json({ error: "User not found" });
             return;
         }
-        const index = user.liked.indexOf(post_id.toString());
+        const postObjectId = new mongoose_1.default.Types.ObjectId(post_id);
+        const index = user.liked.indexOf(postObjectId);
         if (index !== -1) {
             user.liked.splice(index, 1);
             post.likes--;
+            author.totalLikes--;
+            post.likesList = post.likesList.filter((id) => id.toString() !== user_id);
         }
         else {
-            user.liked.push(post_id.toString());
+            user.liked.push(postObjectId);
             post.likes++;
+            author.totalLikes++;
+            post.likesList.push(user_id);
         }
-        yield user.save();
-        yield post.save();
-        res.status(200).json({ message: "Post liked/unliked successfully", post });
+        yield Promise.all([user.save(), post.save(), author.save()]);
+        res.status(200).json({ message: "Post liked/unliked successfully" });
     }
     catch (error) {
         console.error("Error liking post:", error);
